README: mention how to develop on (pure) Windows
The highly recommended paved path for developing this site on Windows is
to use the Windows Subsystem for Linux. The reason for that is that
Windows' filename restrictions prevent Hugo from accommodating the URLs
containing URL-encoded question marks that git-scm.com has for
historical reasons.

However, I understand that there _are_ circumstances under which it
might be necessary to develop this site in pure Windows, without the
help of Linux.

Let's describe what it takes to do that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>