search: initialize the results URL earlier
When a person types really fast into the search box and also hits the
Enter key really quickly after that, the current code directs the
browser to https://git-scm.com/undefined.

The reason? At that stage, the search results had not yet been retrieved
and as a consequence, the list of results to be shown in the live search
were not populated. But that's exactly from where this URL was obtained.

Let's just initialize that URL a lot earlier.

This fixes https://github.com/git/git-scm.com/issues/2030

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>